# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DropFlow - Dropshipping Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers/find:
    post:
      operationId: findSuppliers
      tags: [finder]
      summary: Find and rank suppliers for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierFinderInput"
      responses:
        "200":
          description: Supplier finder result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierFinderResult"
...
  /suppliers/find/history/{id}:
    delete:
      operationId: deleteSupplierFinderResult
      tags: [finder]
      summary: Delete a finder result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- RESEARCH ----------
  /research/analyze:
    post:
      operationId: analyzeProduct
      tags: [research]
      summary: AI-analyze a product keyword or URL for market research
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchInput"
      responses:
        "200":
          description: Research report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
...
  /research/history/{id}:
    delete:
      operationId: deleteResearchReport
      tags: [research]
      summary: Delete a research report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- DASHBOARD ----------
...
components:
  schemas:
...
    SupplierFinderInput:
      type: object
      required: [productName]
      properties:
        productName:
          type: string
          minLength: 1
          description: Product name to find suppliers for
        targetCostPrice:
          type: number
          description: Optional target cost price budget
        preferredCountry:
          type: string
          description: Optional preferred supplier country
    SupplierMatch:
      type: object
      required:
        [
          name,
          country,
          estimatedCostPrice,
          shippingDays,
          rating,
          matchScore,
          matchReason,
          pros,
          isExisting,
        ]
      properties:
        supplierId:
          type: ["integer", "null"]
        name:
          type: string
        country:
          type: string
        estimatedCostPrice:
          type: number
        shippingDays:
          type: integer
        rating:
          type: number
        matchScore:
          type: integer
        matchReason:
          type: string
        pros:
          type: array
          items:
            type: string
        isExisting:
          type: boolean
        website:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
    SupplierFinderResult:
      type: object
      required: [id, productName, topPick, matches, sourcingTips, createdAt]
      properties:
        id:
          type: integer
        productName:
          type: string
        targetCostPrice:
          type: ["number", "null"]
        preferredCountry:
          type: ["string", "null"]
        topPick:
          type: string
        matches:
          type: array
          items:
            $ref: "#/components/schemas/SupplierMatch"
        sourcingTips:
          type: array
          items:
            type: string
        createdAt:
          type: string

    # ---- Research ----
    ResearchInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          description: Product keyword or URL to research
    ResearchNiche:
      type: object
      required: [name, score]
      properties:
        name:
          type: string
        score:
          type: integer
    ResearchReport:
      type: object
      required:
        [
          id,
          query,
          demandScore,
          competitionLevel,
          suggestedPrice,
          estimatedMargin,
          topNiches,
          pros,
          cons,
          verdict,
          createdAt,
        ]
      properties:
        id:
          type: integer
        query:
          type: string
        demandScore:
          type: integer
          description: 0-100 demand score
        competitionLevel:
          type: string
          enum: [low, medium, high, very-high]
        suggestedPrice:
          type: number
        estimatedMargin:
          type: integer
          description: Estimated margin percentage
        topNiches:
          type: array
          items:
            $ref: "#/components/schemas/ResearchNiche"
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        verdict:
          type: string
          enum: [strong-buy, buy, hold, avoid]
        summary:
          type: ["string", "null"]
        createdAt:
          type: string

    # ---- Stock Alert ----
...